CShader: add shader source to log if it failed to compile
This has been an issue for a while because we manipulate the
shader source by inserting and appending to it. This creates an
issue where the error line thrown by the glsl compilation doesn't
match what is in the source file. With this we can clearly see the
line that is provided by the error.