Revert some logic
The reason we don't test renderers is because they might not be created
yet. Buffer pools, on the other hand, are created when RPRenderManager is
created, so we ask them about support. Buffer pools, in turn, query either
static functions of the renderer, or process info (which is also created
upon startup), or local properties of the buffer pool.