Simplify shader construction
The lambda approach succeeds in reducing the amount of code, but adds
complexity. In this case, because it's only a few lines, it would be
better to simply duplicate the code to make the code much more
understandable. We're not writing code for the computer, we're writing
code for people. We want the code to be as straight-forward as possible,
and this goal trumps saving a few lines.

That comment withstanding, we could move handling of the default scaling
method to another place and greatly simplify this function.